wordcount intial draft
